cookie / session / login added

+++ controllers/comment.js
@@ -5,21 +5,30 @@ const Comment = require("../models/comment");
 exports.postCommentProcess = (req, res, next) => {
   const postId = req.params.postId;
   const comment = req.body.comment;
-
+  const userId = req.session.userId;
   Comment.create({
     content: comment,
     author: "Kim",
-    userId: 1,
+    userId: userId,
     postId: postId,
   });
   res.redirect(`/post/postId=${postId}`);
 };
-exports.deleteCommentProcess = (req, res, next) => {
+exports.deleteCommentProcess = async (req, res, next) => {
   const postId = req.params.postId;
   const commentId = req.params.commentId;
+  const userId = req.session.userId;
 
-  Comment.destroy({
-    where: { id: commentId },
-  });
+  await User.findOne({ commentId: commentId })
+    .then((user) => {
+      if (userId === user.id) {
+        Comment.destroy({
+          where: { id: commentId },
+        });
+      }
+    })
+    .catch((err) => {
+      console.log(err);
+    });
   res.redirect(`/post/postId=${postId}`);
 };
